Also convert acceleration_rate to uint32_t
acceleration_rate is also unsigned

+++ Firmware/planner.cpp
@@ -1132,7 +1132,7 @@ Having the real displacement of the head, we can calculate the total movement le
     block->acceleration_st = (block->acceleration_st + (bresenham_oversample >> 1)) / bresenham_oversample;
 #endif
 
-  block->acceleration_rate = (long)((float)block->acceleration_st * (16777216.0 / (F_CPU / 8.0)));
+  block->acceleration_rate = ((float)block->acceleration_st * (16777216.0 / (F_CPU / 8.0)));
 
 #ifdef LIN_ADVANCE
   if (block->use_advance_lead) {
